Further language data cleanup and README enhancements
