+++ src/main.ts
@@ -1,12 +1,12 @@
 import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
-import { AllLanguageData } from './constants/language-data';
+import { AllLanguageData, Locales } from './constants/languages';
 
 import { createI18n } from 'vue-i18n'
 
 const i18n = createI18n({
-  locale: 'en',
+  locale: Locales.English,
   messages: AllLanguageData,
 });
